rename functions and more
