+++ test_util/src/servers/mod.rs
@@ -11,7 +11,6 @@ use denokv_proto::datapath::ReadRangeOutput;
 use denokv_proto::datapath::SnapshotRead;
 use denokv_proto::datapath::SnapshotReadOutput;
 use denokv_proto::datapath::SnapshotReadStatus;
-use futures::Future;
 use futures::FutureExt;
 use futures::Stream;
 use futures::StreamExt;
@@ -26,27 +25,29 @@ use http_body_util::BodyExt;
 use http_body_util::Empty;
 use http_body_util::Full;
 use hyper1 as hyper;
-use hyper_util::rt::TokioIo;
 use pretty_assertions::assert_eq;
 use prost::Message;
 use std::collections::HashMap;
 use std::convert::Infallible;
 use std::env;
-use std::io;
 use std::net::Ipv6Addr;
 use std::net::SocketAddr;
 use std::net::SocketAddrV6;
 use std::path::PathBuf;
-use std::pin::Pin;
 use std::result::Result;
 use std::time::Duration;
 use tokio::io::AsyncWriteExt;
-use tokio::net::TcpListener;
 use tokio::net::TcpStream;
 
 mod grpc;
+mod hyper_utils;
 mod registry;
 mod ws;
+
+use hyper_utils::run_server;
+use hyper_utils::run_server_with_acceptor;
+use hyper_utils::ServerKind;
+use hyper_utils::ServerOptions;
 
 use super::https::get_tls_listener_stream;
 use super::https::SupportedHttpVersions;
@@ -94,8 +95,6 @@ pub async fn run_all_servers() {
     return hyper_hello(port.parse::<u16>().unwrap()).await;
   }
 
-  // TODO(bartlomieju): in a follow up all these `wrap_` handlers could be removed
-  // in favor of spawning a hyper server directly with a config object
   let redirect_server_fut = wrap_redirect_server(REDIRECT_PORT);
   let double_redirects_server_fut =
     wrap_double_redirect_server(DOUBLE_REDIRECTS_PORT);
@@ -185,7 +184,7 @@ async fn hyper_hello(port: u16) {
       http_body_util::Full::new(Bytes::from("Hello World!")),
     )))
   };
-  run_hyper1_server(
+  run_server(
     ServerOptions {
       addr,
       error_msg: "server error",
@@ -1167,152 +1166,9 @@ async fn download_npm_registry_file(
   Ok(())
 }
 
-#[derive(Clone)]
-struct DenoUnsyncExecutor;
-
-impl<Fut> hyper::rt::Executor<Fut> for DenoUnsyncExecutor
-where
-  Fut: Future + 'static,
-  Fut::Output: 'static,
-{
-  fn execute(&self, fut: Fut) {
-    deno_unsync::spawn(fut);
-  }
-}
-
-#[derive(Debug, Clone, Copy)]
-enum ServerKind {
-  Auto,
-  OnlyHttp1,
-  OnlyHttp2,
-}
-
-async fn hyper1_serve_connection<I, F, S>(
-  io: I,
-  service_fn_handler: F,
-  error_msg: &'static str,
-  kind: ServerKind,
-) where
-  I: hyper::rt::Read + hyper::rt::Write + Unpin + 'static,
-  F: Fn(Request<hyper::body::Incoming>) -> S + Copy + 'static,
-  S: Future<
-      Output = Result<
-        Response<UnsyncBoxBody<Bytes, Infallible>>,
-        anyhow::Error,
-      >,
-    > + 'static,
-{
-  let service = hyper1::service::service_fn(service_fn_handler);
-
-  let result: Result<(), anyhow::Error> = match kind {
-    ServerKind::Auto => {
-      let builder =
-        hyper_util::server::conn::auto::Builder::new(DenoUnsyncExecutor);
-      builder
-        .serve_connection(io, service)
-        .await
-        .map_err(|e| anyhow::anyhow!("{}", e))
-    }
-    ServerKind::OnlyHttp1 => {
-      let builder = hyper1::server::conn::http1::Builder::new();
-      builder
-        .serve_connection(io, service)
-        .await
-        .map_err(|e| e.into())
-    }
-    ServerKind::OnlyHttp2 => {
-      let builder =
-        hyper1::server::conn::http2::Builder::new(DenoUnsyncExecutor);
-      builder
-        .serve_connection(io, service)
-        .await
-        .map_err(|e| e.into())
-    }
-  };
-
-  if let Err(e) = result {
-    eprintln!("{}: {:?}", error_msg, e);
-  }
-}
-
-#[derive(Debug, Clone, Copy)]
-struct ServerOptions {
-  error_msg: &'static str,
-  addr: SocketAddr,
-  kind: ServerKind,
-}
-
-async fn run_hyper1_server<F, S>(options: ServerOptions, service_fn_handler: F)
-where
-  F: Fn(Request<hyper::body::Incoming>) -> S + Copy + 'static,
-  S: Future<
-      Output = Result<
-        Response<UnsyncBoxBody<Bytes, Infallible>>,
-        anyhow::Error,
-      >,
-    > + 'static,
-{
-  let fut: Pin<Box<dyn Future<Output = Result<(), anyhow::Error>>>> =
-    async move {
-      let listener = TcpListener::bind(options.addr).await?;
-      loop {
-        let (stream, _) = listener.accept().await?;
-        let io = TokioIo::new(stream);
-        deno_unsync::spawn(hyper1_serve_connection(
-          io,
-          service_fn_handler,
-          options.error_msg,
-          options.kind,
-        ));
-      }
-    }
-    .boxed_local();
-
-  if let Err(e) = fut.await {
-    eprintln!("{}: {:?}", options.error_msg, e);
-  }
-}
-
-// TODO(bartlomieju): dedup with `run_hyper1_server_inner`
-async fn run_hyper1_server_with_acceptor<'a, A, F, S>(
-  mut acceptor: Pin<Box<A>>,
-  service_fn_handler: F,
-  error_msg: &'static str,
-  kind: ServerKind,
-) where
-  A: Stream<Item = io::Result<rustls_tokio_stream::TlsStream>> + ?Sized,
-  F: Fn(Request<hyper::body::Incoming>) -> S + Copy + 'static,
-  S: Future<
-      Output = Result<
-        Response<UnsyncBoxBody<Bytes, Infallible>>,
-        anyhow::Error,
-      >,
-    > + 'static,
-{
-  let fut: Pin<Box<dyn Future<Output = Result<(), anyhow::Error>>>> =
-    async move {
-      while let Some(result) = acceptor.next().await {
-        let stream = result?;
-        let io = TokioIo::new(stream);
-        deno_unsync::spawn(hyper1_serve_connection(
-          io,
-          service_fn_handler,
-          error_msg,
-          kind,
-        ));
-      }
-      Ok(())
-    }
-    .boxed_local();
-
-  if let Err(e) = fut.await {
-    eprintln!("{}: {:?}", error_msg, e);
-  }
-}
-
 async fn wrap_redirect_server(port: u16) {
   let redirect_addr = SocketAddr::from(([127, 0, 0, 1], port));
-  run_hyper1_server(
+  run_server(
     ServerOptions {
       addr: redirect_addr,
       error_msg: "Redirect error",
@@ -1325,7 +1181,7 @@ async fn wrap_redirect_server(port: u16) {
 
 async fn wrap_double_redirect_server(port: u16) {
   let double_redirects_addr = SocketAddr::from(([127, 0, 0, 1], port));
-  run_hyper1_server(
+  run_server(
     ServerOptions {
       addr: double_redirects_addr,
       error_msg: "Double redirect error",
@@ -1338,7 +1194,7 @@ async fn wrap_double_redirect_server(port: u16) {
 
 async fn wrap_inf_redirect_server(port: u16) {
   let inf_redirects_addr = SocketAddr::from(([127, 0, 0, 1], port));
-  run_hyper1_server(
+  run_server(
     ServerOptions {
       addr: inf_redirects_addr,
       error_msg: "Inf redirect error",
@@ -1351,7 +1207,7 @@ async fn wrap_inf_redirect_server(port: u16) {
 
 async fn wrap_another_redirect_server(port: u16) {
   let another_redirect_addr = SocketAddr::from(([127, 0, 0, 1], port));
-  run_hyper1_server(
+  run_server(
     ServerOptions {
       addr: another_redirect_addr,
       error_msg: "Another redirect error",
@@ -1364,7 +1220,7 @@ async fn wrap_another_redirect_server(port: u16) {
 
 async fn wrap_auth_redirect_server(port: u16) {
   let auth_redirect_addr = SocketAddr::from(([127, 0, 0, 1], port));
-  run_hyper1_server(
+  run_server(
     ServerOptions {
       addr: auth_redirect_addr,
       error_msg: "Auth redirect error",
@@ -1377,7 +1233,7 @@ async fn wrap_auth_redirect_server(port: u16) {
 
 async fn wrap_basic_auth_redirect_server(port: u16) {
   let basic_auth_redirect_addr = SocketAddr::from(([127, 0, 0, 1], port));
-  run_hyper1_server(
+  run_server(
     ServerOptions {
       addr: basic_auth_redirect_addr,
       error_msg: "Basic auth redirect error",
@@ -1390,7 +1246,7 @@ async fn wrap_basic_auth_redirect_server(port: u16) {
 
 async fn wrap_abs_redirect_server(port: u16) {
   let abs_redirect_addr = SocketAddr::from(([127, 0, 0, 1], port));
-  run_hyper1_server(
+  run_server(
     ServerOptions {
       addr: abs_redirect_addr,
       error_msg: "Absolute redirect error",
@@ -1415,7 +1271,7 @@ async fn wrap_main_ipv6_server(port: u16) {
 }
 
 async fn wrap_main_server_for_addr(main_server_addr: &SocketAddr) {
-  run_hyper1_server(
+  run_server(
     ServerOptions {
       addr: *main_server_addr,
       kind: ServerKind::Auto,
@@ -1429,7 +1285,7 @@ async fn wrap_main_server_for_addr(main_server_addr: &SocketAddr) {
 async fn wrap_main_https_server(port: u16) {
   let tls = get_tls_listener_stream("https", port, Default::default()).await;
   let tls_acceptor = tls.boxed_local();
-  run_hyper1_server_with_acceptor(
+  run_server_with_acceptor(
     tls_acceptor,
     main_server,
     "HTTPS server error",
@@ -1446,7 +1302,7 @@ async fn wrap_https_h1_only_tls_server(port: u16) {
   )
   .await;
 
-  run_hyper1_server_with_acceptor(
+  run_server_with_acceptor(
     tls.boxed_local(),
     main_server,
     "HTTP1 only TLS server error",
@@ -1463,7 +1319,7 @@ async fn wrap_https_h2_only_tls_server(port: u16) {
   )
   .await;
 
-  run_hyper1_server_with_acceptor(
+  run_server_with_acceptor(
     tls.boxed_local(),
     main_server,
     "HTTP2 only TLS server error",
@@ -1474,7 +1330,7 @@ async fn wrap_https_h2_only_tls_server(port: u16) {
 
 async fn wrap_http_h1_only_server(port: u16) {
   let main_server_http_addr = SocketAddr::from(([127, 0, 0, 1], port));
-  run_hyper1_server(
+  run_server(
     ServerOptions {
       addr: main_server_http_addr,
       error_msg: "HTTP1 only server error:",
@@ -1487,7 +1343,7 @@ async fn wrap_http_h1_only_server(port: u16) {
 
 async fn wrap_http_h2_only_server(port: u16) {
   let main_server_http_addr = SocketAddr::from(([127, 0, 0, 1], port));
-  run_hyper1_server(
+  run_server(
     ServerOptions {
       addr: main_server_http_addr,
       error_msg: "HTTP1 only server error:",
@@ -1515,7 +1371,7 @@ async fn wrap_client_auth_https_server(port: u16) {
     }
   };
 
-  run_hyper1_server_with_acceptor(
+  run_server_with_acceptor(
     tls.boxed_local(),
     main_server,
     "Auth TLS server error",
